requirements.txt
absl-py==2.1.0
bleach==6.1.0
certifi==2024.6.2
charset-normalizer==3.3.2
contourpy==1.2.1
cycler==0.12.1
filelock==3.15.4
fonttools==4.53.0
fsspec==2024.6.0
grpcio==1.64.1
idna==3.7
imageio==2.34.1
importlib_metadata==8.0.0
importlib_resources==6.4.0
Jinja2==3.1.4
joblib==1.4.2
kaggle==1.6.14
kiwisolver==1.4.5
lazy_loader==0.4
Markdown==3.6
MarkupSafe==2.1.5
matplotlib==3.9.0
monai==1.3.1
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4
opencv-python==4.10.0.82
packaging==24.1
pandas==2.2.2
pillow==10.3.0
protobuf==4.25.3
pydicom==2.4.4
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytz==2024.1
requests==2.32.3
scikit-image==0.22.0
scikit-learn==1.5.0
scipy==1.13.1
six==1.16.0
sympy==1.12.1
tensorboard==2.17.0
tensorboard-data-server==0.7.2
text-unidecode==1.3
threadpoolctl==3.5.0
tifffile==2024.5.22
torch==2.2.2
torchvision==0.17.2
tqdm==4.66.4
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.1
webencodings==0.5.1
Werkzeug==3.0.3
zipp==3.19.2
